chore: move transaction and common component
