+++ src/modules/auth/components/hwWalletLogin/loading.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import routes from 'src/routes/routes';
-import { TertiaryButton } from 'src/theme/buttons';
+import { TertiaryButton } from '@theme/buttons';
 import LoadingIcon from './loadingIcon';
 
 class Loading extends React.Component {
