+++ src/modules/auth/components/hwWalletLogin/loadingIcon.css
