+++ src/modules/common/components/deviceDisconnectDialog/styles.css
