+++ src/modules/common/components/infoBanner/infoBanner.css
@@ -1,4 +1,4 @@
-@import '../../../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .infoBanner {
   background-color: var(--color-strong-ink-blue);
