+++ src/modules/common/components/infoBanner/infoBanner.test.js
