+++ src/modules/common/components/newsFeed/index.js
