+++ src/modules/common/components/newsFeed/news.css
@@ -1,4 +1,4 @@
-@import '../../../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .news {
   text-decoration: none;
