+++ src/modules/common/components/newsFeed/news.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import moment from 'moment';
-import Icon from 'src/theme/Icon';
+import Icon from '@theme/Icon';
 import NewsParser from './newsParser';
 import styles from './news.css';
