+++ src/modules/common/components/newsFeed/newsFeed.test.js
