+++ src/modules/common/components/newsFeed/newsParser.js
