+++ src/modules/common/components/newsFeed/newsParser.test.js
