+++ src/modules/common/components/onboarding/onboarding.css
@@ -1,4 +1,4 @@
-@import '../../../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .onboarding {
   background-color: var(--color-onboarding-bg);
