+++ src/modules/common/components/onboarding/onboarding.test.js
