+++ src/modules/transaction/components/Overview/Overview.css
@@ -1,4 +1,4 @@
-@import '../../../../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .wrapper {
   & .content {
