+++ src/routes/routesMap.js
@@ -2,14 +2,14 @@ import AddBookmark from '@bookmark/manager/AddBookmarkManager';
 import BlockDetails from '@block/manager/blockDetailsManager';
 import Blocks from '@block/manager/blocksManager';
 import Bookmarks from '@bookmark/manager/BookmarkListManager';
-import Dashboard from '@screens/managers/dashboard';
+import Dashboard from 'src/modules/common/components/dashboard';
 import DelegatesMonitor from '@dpos/validator/components/DelegatesMonitorView';
 import DelegatesPerformanceModal from '@dpos/validator/components/delegatePerformanceModal';
-import HwWalletLogin from '@screens/managers/hwWalletLogin';
+import HwWalletLogin from '@auth/components/hwWalletLogin';
 import Login from '@auth/components/Signin';
 import MonitorWallets from '@account/manager/AccountsManager';
 import MonitorNetwork from '@network/manager/networkManager';
-import MonitorTransactions from '@screens/managers/monitor/transactions';
+import MonitorTransactions from '@transaction/components/Transactions';
 import Register from '@auth/manager/SignupManager';
 import RegisterDelegate from '@dpos/validator/manager/registerDelegateManager';
 import Send from '@token/fungible/components/SendView';
@@ -24,7 +24,7 @@ import Request from '@wallet/components/request';
 import UnlockBalanceView from '@dpos/validator/components/UnlockBalanceView';
 import EditVote from '@dpos/validator/manager/editVoteManager';
 import VotingQueue from '@dpos/validator/manager/votingQueueManager';
-import DeviceDisconnect from '@screens/managers/deviceDisconnectDialog';
+import DeviceDisconnect from 'src/modules/common/components/deviceDisconnectDialog';
 import NewReleaseDialog from '@update/detail/info/newReleaseDialog';
 import SearchBar from '@search/components/SearchBar';
 import ReclaimBalance from '@legacy/manager/reclaimBalance';
